Read the AD paper and apply the algorithmic eureka step
That's the step from where SPJ's talk left off (which corresponds
to a portion of the monadic translation) and the non-mutable
"Finally efficient: A Monadic Translation" from the paper.
I haven't yet read the subsequent "Using an Array" section
and the "Implementation" chapter.

This is over twice faster on the simple (but inflated) tests, longest
of which now takes 0.36s instead of 1.24s. The results are fractionally
different (only in the XOR neutral network tests), due to different numerical
errors (so the order of arithmetic operations must be different,
which is somewhat worrying, because this may affect performance,
but I don't understand it nor can I easily control it).