Overhaul Delta with a class, based on Richard's advice
Now the definition does not confuse parameterization by an underlying
scalar and by the rank of the tensor, which rules out all the
impossible pattern matches in buildVectors and opens the way to other
clean-ups and stricter typing. E.g., we can now express in a type
signature whether a function only works on (arbitrary) scalars
or on totally arbitrary tensors.