Add the derivative definition vs perturbation sanity check suggested by Andrew
It's added only for one way of computing forward derivative,
but it extends to the other way and to gradients by transitivity
using other equations tested in the properties. Also, in case
of the convolution, it's only tested for the the primitive convolution,
not the one constructed from other primitive operations,
but again, transitivity extends it, with twice the permissible error.