Optimize computing the value of a function; fixes #11
This is a tiny bit slower than what GHC optimzed the previous
naive code to, but this is not fragile and works even without
SPECIALIZE or INLINABLE pragmas (with fixed HEAD that doesn't
miss specializations by unboxing dictionaries).

BTW, strangely, removing SPECIALIZE from the three most
relevant functions speeds up BENCH1 by 20%. That's significant
for a code that is not fragile (though it does repeatedly place ()
in a second component of a pair, so perhaps GHC sometimes can
optimize this away, but not always).