Remove an optimization that no longer works after the refactoring
This was a crucial optimization, giving many-fold performance gains.
But I no longer understand why it was not invalidating sharing
(or even producing wrong results) previously.
Or perhaps it was evaluating some delta expressions twice,
but not in a exponential way and it was overshadowed by performance gains?
I can't produce tests that make it behave poorly (except I've just
noticed test blowupMult fails a little when run in parallel with others,
with the old code, so it's affected by counter values, so something
may be unsound, not just slow due to broken sharing, but OTOH,
it's not computationally stable, so the optimization firing or not
due to counters, has the right to affect it and indeed it produces
a different value when the optimization is just turned off in the old code).
I don't understand why after the flattened Delta refactoring
it breaks sharing, causing runtime blowup (but not RAM blowup somehow),
and why it doesn't break test results. I think I understand
why preformance is better than previously even without this optimizations
(it prevented allocating for spurious Lets that now are not created
to begin with).