Introduce a fake scalar Ast that has r and Int variables and builds AST
The goal is to inspect functions built using ADVal operations,
in order to efficiently implement differentiation of build and map.
This should also make it easier to generate MLIR (or OpenCL) from deltas,
regardless of build and map.

It may be not far from working, but it chokes on Numeric (Ast r).
To fix this, we'd need to parameterize everything (e.g., Delta0) not only
over scalar r, but also over vector v, matrix m, tensors x and t.
Should be doable and should not bubble up to the user API.
Then we'd have Ast (Vector r) instead of the absurd Vector (Ast r).
By chance that would also introduce variables ranging over vectors, etc.

Or we might produce dummy dual parts for the dual numbers with (Ast r)
scalar, because we are going to analyze the function (recorded
in the primal part), not the delta expression.