Introduce a fake scalar that builds AST with r, Vector and Int variables
The goal is to inspect functions built using ADVal operations,
in order to efficiently implement differentiation of build and map.
If a function doesn't have a good form, there's a fallback
to element-wise build and map, as in a straightforward extension
of the POPL paper (or we could fall back to functions stored on tape,
as in the current fastest implementation we have).

Similar mechanism, but for Delta expressions, could work to generate
MLIR (or OpenCL) from deltas, as soon as we manage to eliminate
arbitrary functions stored on tape by the current build1.

This type-checks and will probably work when all missing
definitions are written and a few thinkos fixed.
To avoid code duplication, this will require full "sufficiently
polymorphic" programming style, that is, type classes VectorLike,
MatrixLike, TensorXLike and TensorSLike. Otherwise all primitive
and user-defined operations will require Vector and Ast(Vector)
variants, etc.

BTW, this is no longer a fake scalar, but also fake Vector and will
readily generalize to other ranks.