Introduce a fake scalar Ast that has r and Int variables and builds AST
The goal is to inspect functions built using ADVal operations,
in order to efficiently implement differentiation of build and map.
Similar mechanism, but for Delta expressions, could work to generate
MLIR (or OpenCL) from deltas, as soon as we manage to eliminate
arbitrary functions stored on tape by the current build1.

This type-checks and will probably work when all missing
definitions are written and a few thinkos fixed.
We have half of the needed expressivity: when inside the function
argument of build or map we can use all values from the context,
but of the newly created values only the scalar ones may depend
on the parameter of the function. Full expressivity may or may not
require a huge rewrite.